@@ -13,31 +13,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.2.2",
-    "@angular/core": "~13.2.2",
-    "@angular/forms": "~13.2.2",
-    "@angular/platform-browser": "~13.2.2",
-    "@angular/platform-browser-dynamic": "~13.2.2",
-    "@angular/router": "~13.2.2",
-    "@capacitor-community/barcode-scanner": "^2.1.0",
+    "@angular/common": "^15.0.0",
+    "@angular/core": "^15.0.0",
+    "@angular/forms": "15.2.10",
+    "@angular/platform-browser": "^15.0.0",
+    "@angular/platform-browser-dynamic": "^15.0.0",
+    "@angular/router": "^15.0.0",
+    "@capacitor-community/barcode-scanner": "^3.0.3",
     "@capacitor/android": "^4.0.0",
     "@capacitor/app": "^4.0.0",
     "@capacitor/core": "^4.0.0",
     "@capacitor/device": "^4.0.0",
     "@capacitor/haptics": "^4.0.0",
-    "@capacitor/ios": "^4.0.0",
+    "@capacitor/ios": "4.7.1",
     "@capacitor/keyboard": "^4.0.0",
     "@capacitor/network": "^4.0.0",
     "@capacitor/preferences": "^4.0.2",
     "@capacitor/splash-screen": "^4.0.0",
     "@capacitor/status-bar": "^4.0.0",
-    "@capacitor/storage": "^1.2.5",
-    "@ionic/angular": "^6.0.0",
-    "angularx-qrcode": "^13.0.15",
+    "@ionic/angular": "^6.7.5",
+    "angularx-qrcode": "^15.0.1",
     "flag-icons": "^6.1.1",
     "hash-it": "^5.0.2",
     "hls.js": "^1.1.5",
-    "ionic-rating-component": "^1.2.1",
+    "ion-rating-stars": "^1.2.4",
     "moment": "^2.29.2",
     "onesignal-cordova-plugin": "^3.0.2",
     "rxjs": "~6.6.0",
@@ -46,15 +45,15 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.3",
-    "@angular-eslint/builder": "~13.0.1",
-    "@angular-eslint/eslint-plugin": "~13.0.1",
-    "@angular-eslint/eslint-plugin-template": "~13.0.1",
-    "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.2.3",
-    "@angular/compiler": "~13.2.2",
-    "@angular/compiler-cli": "~13.2.2",
-    "@angular/language-service": "~13.2.2",
+    "@angular-devkit/build-angular": "^15.2.10",
+    "@angular-eslint/builder": "^15.0.0",
+    "@angular-eslint/eslint-plugin": "^15.0.0",
+    "@angular-eslint/eslint-plugin-template": "^15.0.0",
+    "@angular-eslint/template-parser": "^15.0.0",
+    "@angular/cli": "^15.0.0",
+    "@angular/compiler": "^15.0.0",
+    "@angular/compiler-cli": "^15.0.0",
+    "@angular/language-service": "^15.0.0",
     "@capacitor/assets": "^2.0.4",
     "@capacitor/cli": "^4.0.0",
     "@ionic/angular-toolkit": "^6.0.0",
@@ -77,7 +76,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.4.4"
+    "typescript": "4.8.2"
   },
   "description": "An Ionic project"
 }
