@@ -3,7 +3,6 @@ import {BrowserModule} from '@angular/platform-browser';
 import {RouteReuseStrategy} from '@angular/router';
 import {HttpClientModule} from '@angular/common/http';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import { IonicRatingComponentModule } from 'ionic-rating-component';
 import { QRCodeModule } from 'angularx-qrcode';
 
 import {IonicModule, IonicRouteStrategy} from '@ionic/angular';
@@ -14,7 +13,7 @@ import {AppRoutingModule} from './app-routing.module';
 @NgModule({
   declarations: [AppComponent],
   entryComponents: [],
-  imports: [BrowserModule, IonicModule.forRoot(), AppRoutingModule, HttpClientModule, ReactiveFormsModule, FormsModule, IonicRatingComponentModule, QRCodeModule],
+  imports: [BrowserModule, IonicModule.forRoot(), AppRoutingModule, HttpClientModule, ReactiveFormsModule, FormsModule, QRCodeModule],
   providers: [{provide: RouteReuseStrategy, useClass: IonicRouteStrategy}],
   bootstrap: [AppComponent],
 })
