@@ -19,7 +19,7 @@ export class TimetablePage implements OnInit {
 
     activatedRoute.queryParams.subscribe(params => {
       if (params.dataUrl === undefined) {
-        this.url = 'https://2023.revision-party.net/storage/timetable.json';
+        this.url = 'https://2024.revision-party.net/storage/timetable.json';
       } else {
         this.url = params.dataUrl;
       }
