@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Storage } from '@capacitor/storage';
+import { Preferences } from '@capacitor/preferences';
 
 @Injectable({
   providedIn: 'root'
@@ -7,19 +7,19 @@ import { Storage } from '@capacitor/storage';
 export class StorageService {
 
   async get(key): Promise<any> {
-    const ret = await Storage.get({ key });
+    const ret = await Preferences.get({ key });
     return JSON.parse(ret.value);
   }
 
   async set(key, value): Promise<void> {
-    await Storage.set({ key, value: JSON.stringify(value) });
+    await Preferences.set({ key, value: JSON.stringify(value) });
   }
 
   async remove(key): Promise<void> {
-    await Storage.remove({ key });
+    await Preferences.remove({ key });
   }
 
   async clear() {
-    await Storage.clear();
+    await Preferences.clear();
   }
 }
