@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
   dtToken: 'd711aad8eb1c62702653228e42c99237b90bec3d7902b6b72c7b278154659936',
-  dtUrl: 'https://www.deinetickets.de/api/info/code/rev23/',
+  dtUrl: 'https://www.deinetickets.de/api/info/code/rev24/',
 };
